+++ Changelog.md
@@ -6,6 +6,31 @@ The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
 
+## [3.0.0] - 2024-11-02
+
+### Added
+
+- The new [ReQueue] feature has been added!
+You can now "apply" to more than 5 groups at a time!
+
+How this works:
+
+When you are at the maximum of 5 applications the addon saves data from the group you've tried to apply to.
+Then when one of the 5 groups declines you / gets delisted a popup will show up asking you to refresh the search panel.
+Then the addon asks you if you wanna apply to the next group.
+
+The refresh is only required if you applied to a group since the last refresh, so if you apply to 10 groups and then you just play the game after the first refresh you won't have to refresh anymore.
+While I would like for this to be simpler I have to go through some hoops to get this to work.
+
+These groups will be saved between /reloads or quick relogs.
+Enjoy.
+
+### Known issues
+
+- Sometimes the new [ReQueue] feature says it has another group you can apply to but it actually does not.
+
+
+
 ## [2.9.6a] - 2024-11-01
 
 ### Fixed
@@ -2707,7 +2732,8 @@ Basically just the rating for the bracket and the tier they're currently in
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[unreleased]: https://github.com/NintendoLink07/MythicIOGrabber/compare/2.9.6a..HEAD
+[unreleased]: https://github.com/NintendoLink07/MythicIOGrabber/compare/3.0.0..HEAD
+[3.0.0]: https://github.com/NintendoLink07/MythicIOGrabber/releases/tag/3.0.0
 [2.9.6a]: https://github.com/NintendoLink07/MythicIOGrabber/releases/tag/2.9.6a
 [2.9.6]: https://github.com/NintendoLink07/MythicIOGrabber/releases/tag/2.9.6
 [2.9.5]: https://github.com/NintendoLink07/MythicIOGrabber/releases/tag/2.9.5